Add send cover by file if max size is reached
Telegram's API has a max photo size of 5MB when sending photos by URL.
This behavior caused the bot to not reply to a query made by the user
when the cover image for the movie exceeded the file limit. This patch
solves this issue by fetching the image, storing it into an internal
buffer, downsampling the image by half iteratively until it is smaller
than 5MB, and then sending it as a file instead of as a URL.

This new functionality may cause the bot to halt for a second or two,
since it requires some time to download the image, downsample it until
it is smaller than the required limit, and then upload it to Telegram.
This delay depends on both the host's processing power as well as
network speed.